GraviolaContext init
