+++ frontend/src/main.jsx
@@ -5,11 +5,12 @@ import './index.css'
 import {
     RouterProvider,
 } from "react-router-dom";
-
-
+import App from "./App";
 
 ReactDOM.createRoot(document.getElementById('root')).render(
     <React.StrictMode>
-        <RouterProvider router={router} />
+        <App>
+            <RouterProvider router={router} />
+        </App>
     </React.StrictMode>,
 )
